% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_ammi.R
\name{bayes_ammi}
\alias{bayes_ammi}
\alias{bayes_ammi.default}
\title{Bayesian Estimation of the Additive Main Effects and Multiplicative Interaction Model}
\usage{
\method{bayes_ammi}{default}(.data, .y, .gen, .env, .rep, .nIter)
}
\arguments{
\item{.data}{data.frame}

\item{.y}{Response Variable}

\item{.gen}{Genotypes Factor}

\item{.env}{Environment Factor}

\item{.rep}{Replication Factor}

\item{.nIter}{Number of Iterations}
}
\value{
Genotype by Environment Interaction Model
}
\description{
Performs Bayesian Estimation of the Additive Main Effects and Multiplicative Interaction Model
}
\examples{

data(Maiz)

fm1 <-
  bayes_ammi(
    .data = Maiz,
    .y = Yield,
    .gen = Gen,
    .env = Env,
    .rep = Rep,
    .nIter = 20
  )
names(fm1)
fm1

plot(x = fm1, plot_selection = 1)
plot(x = fm1, plot_selection = 2)
plot(x = fm1, plot_selection = 3)
plot(x = fm1, plot_selection = 4)
plot(x = fm1, plot_selection = 5)
plot(x = fm1, plot_selection = 6)

biplots(fm1)

}
\references{
Crossa, J., Perez-Elizalde, S., Jarquin, D., Cotes, J.M., Viele, K., Liu, G., and Cornelius, P.L. (2011)
 Bayesian Estimation of the Additive Main Effects and Multiplicative Interaction Model
 \emph{Crop Science}, \strong{51}, 1458–1469.
 (\href{https://acsess.onlinelibrary.wiley.com/doi/abs/10.2135/cropsci2010.06.0343}{doi: 10.2135/cropsci2010.06.0343})
}
\author{
\enumerate{
    \item Muhammad Yaseen (\email{myaseen208@gmail.com})
    \item Jose Crossa (\email{j.crossa@cgiar.org})
    \item Sergio Perez-Elizalde (\email{sergiop@colpos.mx})
    \item Diego Jarquin (\email{diego.jarquin@gmail.com})
    \item Jose Miguel Cotes
    \item Kert Viele
    \item Genzhou Liu
    \item Paul L. Cornelius
   }
}
