% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesAMMI2APP.R
\name{BayesAMMI2APP}
\alias{BayesAMMI2APP}
\title{BayesAMMI2APP}
\usage{
BayesAMMI2APP()
}
\value{
output of BayesAMMI2 R package
}
\description{
Launch the BayesAMMI2 Shiny App
}
\examples{
if(interactive()) {
    library(BayesAMMI2)
    BayesAMMI2APP()
  }

}
