% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_eff.R
\name{e_eff}
\alias{e_eff}
\alias{e_eff.default}
\title{Environment Effects}
\usage{
\method{e_eff}{default}(.data, .y, .gen, .env)
}
\arguments{
\item{.data}{A data frame containing the dataset.}

\item{.y}{Response variable (numeric).}

\item{.gen}{Genotype factor variable.}

\item{.env}{Environment factor variable.}
}
\value{
A matrix of environment effects.
}
\description{
Calculates Environment Effects for a given dataset using Genotype, Environment, and Response variables.
}
\examples{
data(Maiz)
e_eff(
  .data = Maiz,
  .y = Yield,
  .gen = Gen,
  .env = Env
)
}
\references{
Crossa, J., Perez-Elizalde, S., Jarquin, D., Cotes, J.M., Viele, K., Liu, G., and Cornelius, P.L. (2011).
Bayesian Estimation of the Additive Main Effects and Multiplicative Interaction Model.
\emph{Crop Science}, \strong{51}, 1458–1469.
(\href{https://acsess.onlinelibrary.wiley.com/doi/abs/10.2135/cropsci2010.06.0343}{doi: 10.2135/cropsci2010.06.0343})
}
\author{
\enumerate{
  \item Muhammad Yaseen (\email{myaseen208@gmail.com})
}
}
