% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.BayesAMMI}
\alias{plot.BayesAMMI}
\title{Plot Results from BayesAMMI Model}
\arguments{
\item{x}{An object of class \code{BayesAMMI} containing results from a Bayesian AMMI analysis.}

\item{plot_selection}{A numeric value (1 to 6) or "all" (default) specifying which plots to generate:
\itemize{
  \item 1: Trace and histogram plots for \eqn{\mu}.
  \item 2: Trace and histogram plots for \eqn{\sigma^2}.
  \item 3: AMMI method biplot.
  \item 4: Bayesian AMMI method biplot.
  \item 5: All trace and histogram plots.
  \item 6: All biplots.
  \item "all": All available plots.
}}

\item{...}{Additional arguments (not used).}
}
\value{
Generates ggplot objects and displays them sequentially.
}
\description{
Generates diagnostic and visualization plots for the BayesAMMI model.
}
\details{
The function generates diagnostic and visualization plots based on the `plot_selection` argument.
}
\examples{
data(Maiz)
fm1 <- bayes_ammi(
    .data = Maiz
 ,  .y    = Yield
 ,  .gen  = Gen
 ,  .env  = Env
 ,  .rep  = Rep
 , .nIter = 20
)
plot(fm1, plot_selection = "all")  # Plot all
plot(fm1, plot_selection = 1)      # Only mu plots


}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
}
